Xic: WindowDesc::WindowID and similar are now all uintptr_t instead of unsigned long.  This fixes a bug in macOS copying from "alt" window, could not pick up objects, size of ID was too small and equality test was failing.  Also (QT), ghosting between windows for the copy from alt window now works,

+++ xic/include/dsp.h
@@ -228,7 +228,7 @@ public:
     int OpenSubwin(const BBox*, WDdbType = WDcddb, const char* = 0,
         bool = false);
     int OpenSubwin(const CDs*, const hyEnt* = 0, bool = false);
-    WindowDesc *Windesc(unsigned long);
+    WindowDesc *Windesc(uintptr_t);
     void RedisplayArea(const BBox*, int = -1);
     void RedisplayAll(int = -1);
     void RedisplayAfterInterrupt();
